Project2/pyproject.toml
[project]
name = "project2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "faiss-cpu>=1.13.2",
    "flask>=3.1.2",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
